13 Nov update; ver 0.5_morpheus
Big update, the database is functional, the whole feedback generation is
there, working (though it still needs some work and much testing). The
pdf code is there, and the change between the data grids  is smooth and
the tags are highlighted. The code is not commented, so you will have to
figure everything out, but it should be quite straight forward. Two
forms added, one additional class and about 10 methods.  The menu form
has two options now (green one was deleted). 2-3 more weeks and we
should be done.
Will you choose the blue button or the red button.